requirements.txt
streamlit>=1.28.0
openai-whisper>=20231117
yt-dlp>=2023.11.16
torch>=2.0.0
torchaudio>=2.0.0
ffmpeg-python>=0.2.0
numpy>=1.24.0
